refactor(clock): extract clock + greeting to js/clock.js
